first working subscriptions
