+++ .babelrc
@@ -3,6 +3,6 @@
   "presets": [
     "react",
     "es2015",
-    "stage-0"
+    "stage-2"
   ]
 }
