+++ data/connectors.js
@@ -5,7 +5,6 @@ var users = [{id: 0, firstName: "Test", lastName: "User"}];
 var id = 1;
 
 function addUser(firstName, lastName) {
-    console.log("Adding " + firstName + " " + lastName);
     return new Promise(resolve => {
         setTimeout(() => {
             let u = {
