+++ data/subscriptions.js
@@ -7,7 +7,13 @@ import schema from './schema';
 const pubsub = new PubSub();
 const subscriptionManager = new SubscriptionManager({
     schema,
-    pubsub
+    pubsub,
+    setupFunctions: {
+        userAdded: (options, args) => ({
+            userAdded: user =>
+            user.firstName === args.firstName,
+        }),
+    },
 });
 
 export { subscriptionManager, pubsub };
