+++ package.json
@@ -7,27 +7,26 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "main": "server.js",
-  "dependencies": {
-    "babel-core": "^6.5.2",
-    "babel-loader": "6.2.3",
-    "babel-polyfill": "6.5.0",
-    "babel-preset-es2015": "6.5.0",
-    "babel-preset-react": "6.5.0",
-    "babel-preset-stage-0": "6.5.0",
-    "body-parser": "^1.15.2",
-    "express": "^4.14.0",
-    "graphql": "^0.8.2",
-    "graphql-server-express": "^0.4.3",
-    "graphql-subscriptions": "^0.2.2",
-    "graphql-tools": "^0.8.1",
-    "subscriptions-transport-ws": "^0.3.1"
-  },
   "devDependencies": {
-    "babel-cli": "6.5.1",
-    "casual": "^1.5.2",
-    "lodash": "^4.11.1",
-    "nodemon": "^1.9.1",
-    "webpack": "^1.13.1"
+    "babel-cli": "6.16.0",
+    "babel-core": "6.17.0",
+    "babel-eslint": "7.1.1",
+    "babel-loader": "6.2.6",
+    "babel-preset-es2015": "6.16.0",
+    "babel-preset-react": "6.16.0",
+    "babel-preset-stage-2": "6.17.0",
+    "babel-register": "6.16.3",
+    "nodemon": "1.11.0"
+  },
+  "dependencies": {
+    "body-parser": "1.15.2",
+    "express": "4.14.0",
+    "graphql": "0.7.2",
+    "graphql-server-express": "^0.4.1",
+    "graphql-subscriptions": "0.2.1",
+    "graphql-tools": "0.8.0",
+    "lodash": "4.16.5",
+    "subscriptions-transport-ws": "0.2.6"
   },
   "author": "Lukas Kugler",
   "license": "MIT"
